Authorization (#3)
* Authorization
1) Assigned user roles
2) Post publish and unpublish

* Authorization
1) Assigned user roles
2) Post publish and unpublish

* Authorization
1) Assigned user roles
2) Post publish and unpublish

* Addressed Authorization feedback

+++ app/controllers/posts_controller.rb
@@ -1,16 +1,20 @@
 # frozen_string_literal: true
 
 class PostsController < ApplicationController
+  before_action :find_post, only: %i[show update edit destroy]
   def index
     @post = Post.all.order('created_at DESC')
   end
 
   def new
     @post = Post.new
+    authorize @post
   end
 
   def create
     @post = Post.new(post_params)
+    @post.user = current_user
+    authorize @post
     if @post.save
       redirect_to @post
     else
@@ -19,11 +23,11 @@ class PostsController < ApplicationController
   end
 
   def show
-    @post = Post.find_by(id: params[:id])
+    authorize @post
   end
 
   def update
-    @post = Post.find_by(id: params[:id])
+    authorize @post
     if @post.update(post_params)
       redirect_to @post
     else
@@ -32,11 +36,11 @@ class PostsController < ApplicationController
   end
 
   def edit
-    @post = Post.find_by(id: params[:id])
+    authorize @post
   end
 
   def destroy
-    @post = Post.find_by(id: params[:id])
+    authorize @post
     flash[:notice] = @post.errors.full_messages.to_sentence unless @post.destroy
     redirect_to posts_path
   end
@@ -44,6 +48,10 @@ class PostsController < ApplicationController
   private
 
   def post_params
-    params.require(:post).permit(:title, :content, :user_id).with_defaults(user_id: current_user.id)
+    params.require(:post).permit(:title, :content, :status, :user_id)
+  end
+
+  def find_post
+    @post = Post.find_by(id: params[:id])
   end
 end
